Extracting attachment logic from newsletters and reusing it for newsletters and resources
