+++ app/models/tipline_newsletter.rb
@@ -13,16 +13,12 @@ class TiplineNewsletter < ApplicationRecord
     3 => 230
   }
 
-  include TiplineNewsletterImage
-  include TiplineNewsletterVideo
-  include TiplineNewsletterAudio
+  include TiplineContentMultimedia
 
   belongs_to :team
   has_many :tipline_newsletter_deliveries
   has_paper_trail on: [:create, :update, :destroy], save_changes: true, ignore: [:updated_at, :created_at], versions: { class_name: 'Version' }, if: proc { |_x| User.current.present? }
   has_shortened_urls
-
-  mount_uploader :header_file, FileUploader
 
   before_validation :set_team, :set_last_scheduled_information
 
@@ -31,32 +27,18 @@ class TiplineNewsletter < ApplicationRecord
   validates_presence_of :time, :timezone
   validates_presence_of :introduction, :team, :language
   validates :introduction, length: { maximum: 180 }
-  validates_presence_of :send_on, if: ->(newsletter) { newsletter.content_type == 'static' }
   validates_format_of :rss_feed_url, with: URI.regexp, if: ->(newsletter) { newsletter.content_type == 'rss' }
-  validates_inclusion_of :number_of_articles, in: 0..3, allow_blank: true, allow_nil: true
-  validates_inclusion_of :language, in: ->(newsletter) { newsletter.team.get_languages.to_a }
-  validates_inclusion_of :header_type, in: ['none', 'link_preview', 'audio', 'video', 'image']
   validates_inclusion_of :content_type, in: ['static', 'rss']
+  validates_inclusion_of :language, in: ->(newsletter) { newsletter.team.get_languages.to_a }
+  validates_presence_of :send_on, if: ->(newsletter) { newsletter.content_type == 'static' }
+  validates_inclusion_of :number_of_articles, in: 0..3, allow_blank: true, allow_nil: true
   validates :first_article, length: { maximum: proc { |newsletter| MAXIMUM_ARTICLE_LENGTH[newsletter.number_of_articles].to_i } }, allow_blank: true, allow_nil: true, if: proc { |newsletter| newsletter.number_of_articles >= 1 }
   validates :second_article, length: { maximum: proc { |newsletter| MAXIMUM_ARTICLE_LENGTH[newsletter.number_of_articles].to_i } }, allow_blank: true, allow_nil: true, if: proc { |newsletter| newsletter.number_of_articles >= 2 }
   validates :third_article, length: { maximum: proc { |newsletter| MAXIMUM_ARTICLE_LENGTH[newsletter.number_of_articles].to_i } }, allow_blank: true, allow_nil: true, if: proc { |newsletter| newsletter.number_of_articles == 3 }
   validate :send_every_is_a_list_of_days_of_the_week
-  validate :header_file_is_supported_by_whatsapp
   validate :not_scheduled_for_the_past
 
   after_save :reschedule_delivery
-  after_commit :convert_header_file, on: [:create, :update]
-
-  # File uploads through GraphQL require this setter
-  # Accepts an array or a single file, but persists only one file
-  def file=(file)
-    @file_set = true
-    self.header_file = [file].flatten.first
-  end
-
-  def new_file_uploaded?
-    !!@file_set
-  end
 
   def parsed_timezone
     timezone = self.timezone
@@ -170,52 +152,8 @@ class TiplineNewsletter < ApplicationRecord
     Bot::Smooch.format_template_message(self.whatsapp_template_name, params, file_url, self.build_content, self.language, file_type, preview_url)
   end
 
-  def self.convert_header_file(id)
-    newsletter = TiplineNewsletter.find(id)
-    new_url = nil
-
-    case newsletter.header_type
-    when 'image'
-      new_url = newsletter.convert_header_file_image
-    when 'video'
-      new_url = newsletter.convert_header_file_video
-    when 'audio'
-      new_url = newsletter.convert_header_file_audio
-    end
-
-    newsletter.update_column(:header_media_url, new_url) unless new_url.nil?
-  end
-
-  # Converts an audio or video file to a video with the specific codecs supported by WhatsApp
-  def convert_header_file_audio_or_video(type)
-    options = nil
-    if type == 'audio'
-      cover_path = File.join(Rails.root, 'public', 'images', 'newsletter_audio_header.png')
-      options = ['-loop', '1', '-i', cover_path, '-c:a', 'aac', '-c:v', 'libx264', '-shortest']
-    elsif type == 'video'
-      options = { video_codec: 'libx264', audio_codec: 'aac' }
-    end
-    url = nil
-    input = nil
-    output_path = nil
-    begin
-      now = Time.now.to_i
-      input = File.open(File.join(Rails.root, 'tmp', "newsletter-#{type}-input-#{self.id}-#{now}"), 'wb')
-      input.puts(URI(self.header_file_url).open.read)
-      input.close
-      output_path = File.join(Rails.root, 'tmp', "newsletter-#{type}-output-#{self.id}-#{now}.mp4")
-      video = FFMPEG::Movie.new(input.path)
-      video.transcode(output_path, options)
-      path = "newsletter/video/newsletter-#{type}-#{self.id}-#{now}"
-      CheckS3.write(path, 'video/mp4', File.read(output_path))
-      url = CheckS3.public_url(path)
-    rescue StandardError => e
-      CheckSentry.notify(e)
-    ensure
-      FileUtils.rm_f input.path
-      FileUtils.rm_f output_path
-    end
-    url
+  def self.content_name
+    'newsletter'
   end
 
   private
@@ -246,32 +184,6 @@ class TiplineNewsletter < ApplicationRecord
       self.last_scheduled_at = Time.now
       self.last_delivery_error = nil
     end
-  end
-
-  def header_file_is_supported_by_whatsapp
-    if self.header_type && self.new_file_uploaded?
-      case self.header_type
-      when 'image'
-        self.validate_header_file_image
-      when 'video'
-        self.validate_header_file_video
-      when 'audio'
-        self.validate_header_file_audio
-      end
-    end
-  end
-
-  def convert_header_file
-    if self.should_convert_header_image? || self.should_convert_header_video? || self.should_convert_header_audio?
-      self.class.delay_for(1.second, retry: 3).convert_header_file(self.id)
-    end
-  end
-
-  def validate_header_file(max_size, allowed_types, message)
-    size_in_mb = (self.header_file.file.size.to_f / (1000 * 1000))
-    type = self.header_file.file.extension.downcase
-    errors.add(:base, I18n.t(message, **{ max_size: "#{max_size}MB" })) if size_in_mb > max_size.to_f
-    errors.add(:header_file, I18n.t('errors.messages.extension_white_list_error', **{ extension: type, allowed_types: allowed_types.join(', ') })) unless allowed_types.include?(type)
   end
 
   def not_scheduled_for_the_past
