+++ public/tipline-content-template.html
@@ -70,7 +70,7 @@
       }
     </style>
   </head>
-  <body class="newsletter">
+  <body class="content">
     <div id="frame">
       <div id="image">
         <div id="text">
